Modify install.sh to download jubatus-mpio and jubatus-msgpack-rpc from GitHub.
The aim of this change is to install develop versions of jubatus-mpio, jubatus-msgpack-rpc, and jubatus_core.
The reason to download from GitHub is that develop versions of them are only on GitHub.
Actual implementation of the installation function are on a subsequent commit.

Availability checks for ruby and autotools are required because they are utilized to generate configure scripts.
jubatus-m* codes on GitHub do not include a configure script.